Add Authorizaiton checks in ContestController

+++ OnlineJudge/Controllers/ContestController.cs
@@ -4,11 +4,13 @@ using System.Web.Http;
 using OnlineJudge.FormModels;
 using OnlineJudge.Repository;
 using OnlineJudge.ResponseModels;
+using OnlineJudge.Services;
 
 namespace OnlineJudge.Controllers{
     [RoutePrefix("api/contests")]
     public class ContestController : ApiController{
         private ContestRepository contest_repository;
+        private UserService user_service = new UserService();
 
         public ContestController(){
             contest_repository = new ContestRepository();
@@ -88,6 +90,10 @@ namespace OnlineJudge.Controllers{
         [HttpOptions]
         [Route("create")]
         public IHttpActionResult ContestCreate([FromBody]ContestCreationFormData data){
+            if (!user_service.IsAuthorizedToCreateContest()){
+                return Unauthorized();
+            }
+
             if (RequestUtility.IsPreFlightRequest(Request)){
                 return Ok();
             }
@@ -111,6 +117,10 @@ namespace OnlineJudge.Controllers{
         [HttpPost]
         [Route("{contest_id}/edit")]
         public IHttpActionResult ContestUpdate(int contest_id, [FromBody]ContestCreationFormData data){
+            if (!user_service.IsAuthorizedToEditContest(contest_id)){
+                return Unauthorized();
+            }
+
             if (RequestUtility.IsPreFlightRequest(Request)){
                 return Ok();
             }
@@ -133,6 +143,10 @@ namespace OnlineJudge.Controllers{
         [HttpGet]
         [Route("{contest_id}/problems")]
         public IHttpActionResult ContestProblemList(int contest_id){
+            if (!user_service.IsAuthorizedToViewContestProblems(contest_id)){
+                return Unauthorized();
+            }
+
             try{
                 var contest = contest_repository.GetContestById(contest_id);
                 return Ok(new ContestDetailsData(contest));
@@ -145,6 +159,10 @@ namespace OnlineJudge.Controllers{
         [HttpGet]
         [Route("{contest_id}/problems/{problem_no}")]
         public IHttpActionResult ContestProblemDetails(int contest_id, int problem_no){
+            if (!user_service.IsAuthorizedToViewContestProblem(contest_id)){
+                return Unauthorized();
+            }
+
             var conest_prblem = contest_repository.GetContestProblem(contest_id, problem_no);
             return Ok(new ProblemDetails(conest_prblem.Problem));
         }
@@ -209,6 +227,10 @@ namespace OnlineJudge.Controllers{
         [HttpPost]
         [Route("{contest_id}/delete")]
         public IHttpActionResult ContestantDelete(int contest_id){
+            if (!user_service.IsAuthorizedToDeleteContest(contest_id)){
+                return Unauthorized();
+            }
+
             try{
                 contest_repository.DeleteContest(contest_id);
                 return Ok();
@@ -236,6 +258,7 @@ namespace OnlineJudge.Controllers{
 
     [RoutePrefix("api/contests")]
     public class ContestSubmissionController : ApiController{
+        private UserService user_service = new UserService();
         private readonly  ContestSubmissionRepository contest_submission_repository;
 
         public ContestSubmissionController(){
@@ -247,6 +270,10 @@ namespace OnlineJudge.Controllers{
         [HttpPost]
         [Route("{contest_id}/problems/{problem_no}/submit")]
         public IHttpActionResult ContestProblemSubmit(int contest_id, int problem_no, [FromBody] SubmissionFormData submission_data){
+            if (!user_service.IsAuthorizedToSubmitToContest(contest_id)){
+                return Unauthorized();
+            }
+            
             if (RequestUtility.IsPreFlightRequest(Request)){
                 return Ok();
             }
